requirements.txt
flask
line-bot-sdk
openai
pydub
requests